Feat: 회원 탈퇴 기능 추가

+++ backend/src/main/java/com/orleave/repository/LoginTrialRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.orleave.entity.LoginTrial;
 
 @Repository
-public interface LoginTrialRepository extends JpaRepository<LoginTrial, Long> {
+public interface LoginTrialRepository extends JpaRepository<LoginTrial, Integer> {
 
 }
